move icons to ui/icons
