+++ resources/js/components/ui/icons/micro/BriefcaseIcon.vue
