+++ resources/js/components/ui/icons/micro/BuildingOfficeIcon.vue
