+++ resources/js/components/ui/icons/mini/UserGroupIcon.vue
